+++ src/test/java/com/intellij/plugins/haxe/ide/quickfix/unresolved/UnresolvedMethodQuickFixTest.java
@@ -1,0 +1,25 @@
+package com.intellij.plugins.haxe.ide.quickfix.unresolved;
+
+import com.intellij.plugins.haxe.ide.quickfix.HaxeQuickFixTestBase;
+import org.jetbrains.annotations.NotNull;
+
+public class UnresolvedMethodQuickFixTest extends HaxeQuickFixTestBase {
+
+
+  @Override
+  protected @NotNull String getBasePath() {
+    return "/unresolved/method";
+  }
+
+
+  public void testCreateMethodAssign() {
+    doSingleTest("_create_method_assign.hx");
+  }
+
+  public void testCreateMethodInIf() {
+    doSingleTest("_create_method_if.hx");
+  }
+  public void testCreateMethodVoid() {
+    doSingleTest("_create_method_void.hx");
+  }
+}